Refactor analyzer dir
