+++ optd/src/dsl/analyzer/mod.rs
@@ -1,8 +1,6 @@
-pub mod context;
 pub mod errors;
 pub mod from_ast;
 pub mod hir;
 pub mod into_hir;
-pub mod map;
 pub mod semantic_checks;
 pub mod type_checks;
