+++ optd/src/dsl/examples/generics.opt
@@ -3,19 +3,62 @@ data Physical
 data LogicalProperties
 data PhysicalProperties
 
-// Generic function that assumes numeric operations
-fn <T> sum(a: T, b: T): T = a + b  // This assumes T supports addition
-
-// Generic function that assumes comparison
-fn <T> max(a: T, b: T): T = 
-    if a > b then a else b  // This assumes T supports comparison
-
-fn main(): I64 = 
-    let 
-        sum_int = sum(1, 2),
-        sum_float = sum(1.5, 2.5),
-        
-        max_int = max(5, 3),
-        max_float = max(5.5, 3.5)
+// Generic functions with recursive closures
+fn <T, U> memoize(f: T -> U): T -> U =
+    let
+        cache = [] // Simplified cache
     in
-        sum_int + max_int
+        (x: T) -> {
+            // In a real implementation, we would check the cache
+            // and only compute if needed
+            f(x)
+        }
+
+fn <T> (array: [T]) quicksort(cmp: (T, T) -> I64) =
+    match array
+        | [] -> []
+        \ [pivot .. rest] -> {
+            let
+                partition = (arr: [T], pivot: T, cmp: (T, T) -> I64) -> {
+                    let
+                        less = arr.filter(x: T -> cmp(x, pivot) < 0),
+                        greater = arr.filter(x: T -> cmp(x, pivot) >= 0)
+                    in
+                        (less, greater)
+                },
+
+                result = partition(rest, pivot, cmp),
+                less = result#_0,
+                greater = result#_1,
+                sorted_less = less.quicksort(cmp),
+                sorted_greater = greater.quicksort(cmp)
+            in
+                sorted_less ++ [pivot] ++ sorted_greater
+            }
+
+fn <T> (array: [T]) filter(predicate: T -> Bool) = match array
+    | [] -> []
+    \ [x .. xs] ->
+        if predicate(x) then
+            [x] ++ xs.filter(predicate)
+        else
+            xs.filter(predicate)
+
+// Fixed recursive function implementation
+fn factorial(n: I64): I64 = 
+    if n <= 1 then 1 else n * factorial(n - 1)
+
+fn main(): I64 =
+    let
+        numbers = [5, 3, 8, 1, 2, 9, 4, 7, 6],
+        compare = (a: I64, b: I64) -> a - b,
+            
+        // Use regular factorial function instead of the Y-combinator approach
+        memoized_factorial = memoize(factorial),
+            
+        // Sort numbers and filter
+        sorted = numbers.quicksort(compare),
+        filtered = sorted.filter(x: I64 -> x > 2),
+        result = memoized_factorial(7)
+    in
+        result
